==> 02_network.qmd <==
---
title: "Exploring mobility dynamics in FVG - part 2: network"
format:
  pdf:
    toc: true
    number-sections: true
    colorlinks: true
editor: visual
---

## Objective

Mobility patterns are **encoded in tabular form**, where each row
represents a flow from location A to location B, characterized by
categorical variables (inbound or outbound, resident or traveler) and a
quantitative indicator of the number of journeys.

The **objective** of this notebook is to **construct a network** where
nodes represent locations A, B, C, D, etc., and edges encode information
on mobility flows, within a user-specified time frame. The core task is
to find a meaningful definition for **edge weights**, considering the
magnitude of flows between locations.

Finally, we aim to develop a **statistical model** for edge weight
estimation, allowing for both *point estimates* and *interval
estimates*, providing insights into the dispersion of data and its
evolution over time.

The main dataset has been prepared in script 01. In the first part of
this script however we use a smaller test dataset (available in
subfolder `dummy_data`) to clarify the methodology.

::: {.callout-note appearance="simple"}
## coding specifications

The R code utilizes the `readxl` package to import dummy data from Excel
and the `tidyverse` package for efficient data manipulation.
Specifically, `tidyverse`an efficient syntax and verbs like `mutate`,
`rename`, and `select` to clean and preprocess the data, enhanching
readability and facilitating code debugging and reuse. Netrowks are
created and analysed with`igraph` package.
:::

# Methodology

To explain the methodology, we can use a simple example of flows between
two locations A-B, in single day.

```{r warning=FALSE}
library(tidyverse)
library(readxl)
library(igraph)
```

Loading a toy dataset

```{r warning=FALSE}
#read data a
edges_simple <- read_excel('./dummy_data/simple_AB_1day_directed.xlsx') 
head(edges_simple)
```

This data can be used straightforward to build a network, with an edge
for each row, and weight that matches the direction. Here we show how,
and explain why it is not the an optimal way of encoding the
information.

We assume that the edge has a weight that is $w = +n$ if direction is
$out$, and $w = -n$ if direction is $in$.

```{r}

add_edge_labels <- function(g){
    for (i in 1:ecount(g)) {
        edge <- E(g)[i]
        first_node_name <- V(g)[.from(edge)]$name
        second_node_name <- V(g)[.to(edge)]$name
        direction = E(g)[i]$direction
        weight = E(g)[i]$weight
        edge_label <- paste0(first_node_name, 
                             second_node_name, 
                             direction, 
                             "_",
                             weight)
        E(g)$label[i] <- edge_label
    }
    return(g)
}
```

How to build a DIRECTED graph

```{r}

ee <- edges_simple %>% 
    select(resident_in, connected_with, n, direction) %>%
    pivot_wider(names_from = direction, values_from = n, values_fill = 0) %>%
    mutate(weight = (go + back)/2)  %>%
    select(resident_in, connected_with, weight)

g <- igraph::graph_from_data_frame(ee, directed = TRUE)
E(g)$curved <- TRUE

g <-add_edge_labels(g)
print(g)
plot(g, 
     edge.label = E(g)$label, 
     edge.width = E(g)$n/10, 
     vertex.color = 'white', 
     vertex.size = 30)
strength(g)
coreness(g)
AMd <- igraph::as_adj(g,edges = FALSE, type = 'both', sparse = FALSE)
AMd
as_long_data_frame(g)
 
```

matrix representation of the weights ==\> very useful to calculate the
indices of INBOUND and OUTBOUND flows for each location (as sum of
matrix rows or columns)

```{r}
locs <- unique(c(edges_simple$resident_in, edges_simple$connected_with))
# Initialize an empty matrix
matrix <- matrix(0, nrow = length(locs), ncol = length(locs), dimnames = list(locs, locs))

# Loop through edges and populate the matrix
for (i in 1:nrow(edges_simple)) {
  resident_in <- edges_simple$resident_in[i]
  connected_with <- edges_simple$connected_with[i]
  weight <- edges_simple$n[i]
  matrix[resident_in, connected_with] <- matrix[resident_in, connected_with] + weight
}

print(matrix) 
```

A more meaningful choice for weight is mean_flow = mean(in, out), which
produces a single edge per day. Moreover, we can measure the symmetry of
the flows with `net_flow = out-in` and `sk= net / mean_flow`

```{r}
df <- read_excel('./dummy_data/simple_AB_1day_directed.xlsx') 
df

edges_flow <- df %>%
  #group_by(resident_in, connected_with, direction) %>%
  #summarise(n_mean = mean(n)) %>%
  #ungroup() %>%
  pivot_wider(names_from = direction, values_from = n, values_fill = 0) %>%
  mutate(flow_mean = (go + back)/2)  %>%
  mutate(flow_net = back - go) %>%
  mutate(flow_ratio = flow_net / flow_mean)
edges_flow
```

```{r}
g <- graph_from_data_frame(edges_flow, directed = TRUE)
E(g)$weight = E(g)$flow_mean
E(g)$direction = ""
g <-add_edge_labels(g)
plot(g, 
     edge.label = E(g)$label, 
     edge.width = E(g)$weight/10, 
     vertex.color = 'white', 
     vertex.size = 30)
```

# Using the real data

Load data

```{r}
flows <- read_csv('./data/flows.csv') 
ndays = length( unique(flows$day))
print(ndays)
```

Caluclate network edges from flows

```{r}

edges <- flows  %>% rename(weight = flow_mean) %>%
    select(resident_in, connected_with, weight, day, weekday, month)

hist(edges$weight)

```

```{r}
edges %>%
    group_by(resident_in)%>%
    summarize(ff = mean(weight))%>%
    arrange(-ff)
```

```{r}

ee <- edges %>% 
    select(resident_in, connected_with, month, weight)%>%
    group_by(month, resident_in, connected_with)%>%
    summarize(weight = sum(weight), .groups = "drop") %>%
    select(resident_in, connected_with, weight, month)

g3 <- ee %>% 
    filter(month == 3) %>% 
    igraph::graph_from_data_frame(directed = FALSE) %>%
    igraph::simplify()

g6 <- ee %>% 
    filter(month == 6) %>% 
    igraph::graph_from_data_frame(directed = FALSE) %>%
    igraph::simplify()

```

```{r}

ll = layout.graphopt(g3)
plot(g3, layout = ll,
     vertex.label = NA,
     edge.width = log(E(g3)$weight), 
     vertex.size = log(strength(g3)))
plot(g6, layout = ll, 
     vertex.label = NA,
     edge.width = log(E(g6)$weight), 
     vertex.size = log(strength(g6)))

```

check strength and coreness

```{r}

 

g <- ee %>% filter(month == 3) %>%
    igraph::graph_from_data_frame(directed = FALSE) %>%
    igraph::simplify()

data.frame(loc = V(g)$name, 
                 dd = degree(g),
                 ss = round(strength(g),0),
                 kk = coreness(g)) %>%
    arrange(-ss) %>% 
    ggplot(aes(x = dd, y = kk)) + 
    geom_point(aes(size = ss*2), color = 'red',alpha = 0.5)



g <- edges %>% filter(month == 6) %>%
    igraph::graph_from_data_frame(directed = FALSE) %>%
    igraph::simplify()

data.frame(loc = V(g)$name, 
                 dd = degree(g),
                 ss = round(strength(g),0),
                 kk = coreness(g)) %>%
    arrange(-ss) %>% 
    ggplot(aes(x = dd, y = kk)) + 
    geom_point(aes(size = ss*2), color = 'blue',alpha = 0.5)

```

show neighborhood

```{r}
g1 <- g
selected <- which(V(g1)$name == "Muggia")
nbh<-make_ego_graph(
  g1,
  order = 1,
  nodes = selected,
  mode = "all",
  mindist = 0
)[[1]]

#mean((E(g)$weight))
#E(g)$weight[ E(g)$weight < 150] <- 0

lo = layout.graphopt(nbh)
plot(nbh, vertex.size = sqrt(strength(nbh))/10, edge.width = E(nbh)$weight/2000, vertex.label = NA, layout = lo)
plot(nbh, vertex.size = sqrt(strength(nbh))/10, edge.width = E(nbh)$weight/2000,layout = lo)

```

```{r}
file_name = "mobility_fvg_sample_01.graphml"   
g %>% igraph::write_graph(file_name, format="graphml")

     
```

# Exploring network features

to do:

-   distribution of weights

-   centrality measures: degree, strength, coreness

```{r}
V(g)$str<-strength(g)
df<- data.frame(n = V(g)$name, str = round(strength(g),1))
df <- df %>% arrange(-str) 
df %>% head(10)
```

```{r}
plot(g, vertex.size = sqrt(V(g)$str)/10, vertex.label = NA)
```

-   show graph with a layout that is using geographical coordinates LAT
    LON nrmalized to range -1,1

-   show ego-subgraph of some locations

-   igraph plots: reduce the white area around the plot
